feat: 홈화면 1차 완료
- 네이밍 수정 예정
- TODO 처리 예정..
