--- Projects/App/Sources/Home/Past/PastHomePage.swift
@@ -1,21 +1,0 @@
-//
-//  PastHomeView.swift
-//  Lyfe
-//
-//  Created by 박서연 on 2024/02/10.
-//  Copyright © 2024 iOSteam. All rights reserved.
-//
-
-import SwiftUI
-
-struct PastHomePage: View {
-    @State private var postType: HomePicker = .today
-    
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-#Preview {
-    PastHomePage()
-}
