+++ Projects/App/Sources/Home/Picker/HomePickerPage.swift
@@ -8,34 +8,14 @@
 
 import SwiftUI
 
-enum HomePicker: String, CaseIterable {
-    case today = "오늘의 주제"
-    case past = "과거 베스트"
-}
-
 struct HomePickerPage: View {
-    @Binding var selected: HomePicker
-    
     var body: some View {
-        VStack(alignment: .leading) {
-            Picker("주제 선택", selection: $selected) {
-                ForEach(HomePicker.allCases, id: \.self) { selected in
-                    Text(selected.rawValue)
-                }
-            }
-            .pickerStyle(MenuPickerStyle())
-            .tint(Color.black)
-            .onChange(of: selected) { newValue in
-                if newValue == .today {
-                    TodayHomePage()
-                } else {
-                    PastPostPage()
-                }
-            }
+        VStack {
+            
         }
     }
 }
 
 #Preview {
-    HomePickerPage(selected: .constant(.past))
+    HomePickerPage()
 }
